Create a "Menard" shape plasma
Also have an option for a pure-Elliptical Plasma.

These are in elliptical_plasma.[Menard,Elliptical]PlasmaGeometry
This allows easy switching between the two, and they have the same
interface.

The difference between the two is that menard includes a "κ area
fraction" of 0.8 which reduces the cross-sectional area and volume. I
suppose this accounts for the cusped shape of a one- or two-divertor
plasma.