Change 'max' to 'out' and 'min' to 'in' for radii
I want to reserve 'max' and 'min' in variable names for constraint-like
properties, not the actual values of properties. So far the exception is
the Ob TF R_min, the minimum radius for the outer TF leg. It might be
larger, for example with a Princeton Dee coil that requries a certain
interior height.