Initial commit of GeneromakCosting
This implements fairly closely the 2016 Sheffield paper for generomak
costing. The corporate tax, interest, escalation, and inflation rates are (implicitly)
the same as in that paper, and should perhaps be updated via explicit
calculations for F_CRO, tax rates, and the return rates (and fractions)
of debt, stock, and preferred stock involved in financing the project.

There is an option to either use the 'exact' formulation or a slightly
improved formulation. The slight improvements make the cost depend on
the gross electric power, not just the net electric power generated.