Modify bootstrap.py to add a weak δ dependence
Now bootstrap current is proportional to ft / (1 - ft) where ft is the
Sauter formula ε- and δ-dependent trapped particle fraction. The ρ=0.75
flux surface is used to compute these trapped particle fractions. I
assume that δ there is 0.5 of δ of the LCFS.

This change is the first time that δ is used for the plasma model. It is
calibrated to not change for δ=0 (which is assumed for the Menard
model).