Modify PolarParallelCurve to have a radial limiter
This allows the inboard side of the blanket or shield to be pushed out
until it's at a minimum radius. This allows the thickness to poloidally
vary, being a constant except for on the inboard side. The resulting
limiter takes the form of a straight vertical wall on the inboard side.