Generalize the SOL width model
Add a 'Horacek SOL Width' that loads a power law model from
data/fits.yaml. Make the HD model a different choice. Separate the
concept of midplane power width λq from projected-to-midplane width λint
and the final heat flux on target, which is widened by expansion and
tilt factors.

Note that this implies breaking changes for the config tree structure.