Add GenX costing interface
This is an optional 'attachment' to the Generomak costing module which
outputs an interface to GenX for investment cost, fixed O&M cost, and
variable O&M cost. Fuel is nil (is included either in F O&M or in V O&M,
since I do't want to deal with a 'heat rate' input and a separate D
price series, etc)

I added an option for this in the top level of the costing.yaml because
it should be easy/necessary to build such an interface for other costing
codes as well; they should all support this simple (inv, fixed,
variable) interface.