Pass OpenMDAO derivative checking warnings
With the version bump there is a new warning against derivatives being
checked using the same method they are calculated with. To pass this, we
add some analytic derivatives and set others to be checked via fd, or by
fd with central differencing. There's certainly places here that could
use more analytic derivatives.